foreground service

+++ app/src/main/java/com/orvio/app/domain/model/UserStats.kt
@@ -55,5 +55,5 @@ data class KeyStats(
 data class CreditStats(
     val balance: Int,
     val mode: String,
-    val cashbackPoints: Int = 0
+    val cashbackPoints: Float = 0f
 )
